separate test cases & remove dependencies to GuzzleHttp\Psr7\Stream
